refactor: move channels_v2/whatsapp_channel.py to channels/whatsapp_channel.py
